Extract RequestDownloadDialog component with self-contained state and logic
